abi-templater/requirement.txt
# Installing the required python modules

pyyaml
jinja2